MueLu: fix instantiation of MutuallyExclusiveTime
This class seems to be explicitly instantiated
for only two types as 'TagName'.
It also has static member variables.
Because it has static member variables,
either explicit instantiation declarations
must be given (in the case of explicit instantiation),
or the definition of the static member must be
provided in the header file.
I went with explicit instantiation declarations.
In addition, I separated the definitions of the
methods and static variables since they don't
need to be visible anywhere besides the translation
units doing explicit instantiation.
In this case its just one cpp file, so I put the
definitions there.
This is a small victory for reduced compile times
by including less headers and parsing less code.
If instantiations are wanted elsewhere in the future,
the cpp file can be split into a _def.hpp and cpp.
